feat(schema): Expose the Index's Registry Config
